move fastest dataset serialization tests to kernel tests
